refactor(bot): extract bot/tool_executors.py — closes #120
Migrates the 24 ``execute_*`` async tool implementations + the dispatch
table builder out of ``bot/core.py`` into a single-file checkpoint at
``bot/tool_executors.py``. Per-tool split is deferred to a follow-up
(see ADR 0001 "considered options").

Module surface (28 re-exports):

* 24 async tool executors covering the OpenAI function-calling surface:
  ``execute_omicsclaw`` (the largest), ``execute_replot_skill``,
  ``execute_save_file`` / ``execute_write_file`` /
  ``execute_make_directory`` / ``execute_move_file`` /
  ``execute_remove_file`` / ``execute_get_file_size``,
  ``execute_generate_audio``, ``execute_parse_literature``,
  ``execute_fetch_geo_metadata``, ``execute_list_directory``,
  ``execute_inspect_file`` / ``execute_inspect_data``,
  ``execute_remember`` / ``execute_recall`` / ``execute_forget``,
  ``execute_read_knowhow``, ``execute_consult_knowledge``,
  ``execute_resolve_capability``, ``execute_list_skills_in_domain``,
  ``execute_create_omics_skill``, ``execute_web_method_search``,
  ``execute_custom_analysis_execute``.
* Dispatch builder: ``_available_tool_executors`` /
  ``_build_tool_runtime`` / ``get_tool_runtime`` /
  ``get_tool_executors``. The lazy ``bot.core.TOOL_EXECUTORS`` /
  ``TOOL_RUNTIME`` / ``TOOLS`` attributes (resolved via
  ``bot.core.__getattr__``) continue to work because the underlying
  builders re-export through bot.core.

Cross-module access:

* Stable bot.core symbols (path constants, registry, ``audit``,
  ``received_files`` and other in-place-mutated dicts,
  ``get_skill_runner_python``) imported at module top.
* Runtime-reassigned globals (``memory_store``, ``llm``,
  ``OMICSCLAW_MODEL``, ``LLM_PROVIDER_NAME``, ``session_manager``)
  accessed via ``_core.<name>`` at call time so they see the values
  ``bot.session.init()`` writes after the modules finish loading.
  Source code references rewritten to use ``_core.X`` form for those
  five names.
* Sibling helpers (``bot.skill_orchestration`` for ``_lookup_skill_info``
  / ``_format_auto_route_banner`` / ``_classify_env_error`` / ...,
  ``omicsclaw.runtime.preflight.sc_batch`` for
  ``_maybe_require_batch_key_selection`` / ...) imported from canonical
  homes, not via the bot.core re-export — clearer dependency graph.

PR #122 review Important #2 follow-through: ``omicsclaw/runtime/
preflight/sc_batch.py`` now imports ``execute_omicsclaw`` from
``bot.tool_executors`` (was ``bot.core``).

Tests added (``tests/bot/test_tool_executors.py``):

* Identity-comparison contract for all 28 re-exports
  (``bot.core.X is bot.tool_executors.X``).
* Dispatch-table sanity: ``_available_tool_executors()`` lists at
  least the 24 native executors (engineering tools added on top
  bring the total higher).

Verification:

* Bot regression net: 226 + 2 (new file) = 228 passed, 1 skipped, 4
  xfailed — no regressions.
* ``tests/eval/`` + ``tests/test_knowhow.py``: 121 passed, 17
  deselected. Eval baseline 14/15 against ``deepseek-v4-pro`` unaffected.
* ``bot/core.py``: 3191 → 1441 LOC (−1750). Cumulative drop from main
  baseline: 5160 → 1441 (−3719, −72.1%).
* ``bot/tool_executors.py`` lands at 1900 LOC (single-file checkpoint
  per ADR 0001).

Closes #120.
